Add a contact placeholder picture.
Previously, there was no generic contact icon in the framework.
ContactsProvider needs to have a generic contact icon to use
for contacts that don't have photos. Since there are multiple
packages (MyFaves, GoogleSubscribedFeedsProvider) that use
ContactsProvider.java directly, adding the icon in the
ContactsProvider package is not an option.

The icon is a copy of
packages/apps/Contacts/res/drawable-finger/ic_contact_picture.png