Tear down data, if a disconnect request is pending.
If there is a pending disconnect, and the previous connection
attempt succeeds, we don't disconnect and hence the state remains
in DISCONNECTING. Ex: user is on Wifi, location app calls
enableApnType(SUPL) and then we set state to INITING. It then
calls disableApnType(SUPL) we register a pending disconnect.
When the data connection attempt is successful, we don't disconnect,
as the PDP Connection State has not been set to ACTIVE and we check
for this state. This fix removes the dependency on ACTIVE state check.