Fix some 12- vs 24-hour problems in the FormatDateRange strings.
Use a resource instead of a hardcoded string for the 24-hour format
since it is not exactly the same in every locale.

Make sure the 12-hour format is actually for a 12-hour clock, even in
locales where this is not a normal thing to do.  In the cap_ampm version,
do not have it try to capitalize "am" and "pm" if these are non-ASCII
strings, since strftime() doesn't know about Unicode and will mess it up.

Add a comment so that people don't think the YEAR_IN_MILLIS constant is
actually the length of any real year.