Copies the Contacts.Intents class to ContactsContract.
ContactsContract now contains all the useful constants that are
contained in Contacts.Intents.  Because we have to continue to
support the Contacts interface, the constants are kept there,
but now simply reference the real versions in Contacts.Intents.
Contacts.Intents will be deprecated and developers are encouraged
to move to Contacts.Intents.