Update text webcore thread's text generation number.
When the user edits text in a textfield, we increase a generation
number so we can mark changes from webkit to be out of date.  With
this change, update webcore's notion of the text generation number
in deleteSelection and replaceTextfieldText, in addition to
passToJs.  Requires a change in external/webkit.